dnscache: split implementation out of hostip.c
Give the dnscache implementation its own source file, as
hostip.c became unwieldy.

have Curl_dnscache_get() return a CURLcode for negative entries
fix include for unit1305.c
