+++ lib/asyn-ares.c
@@ -322,14 +322,14 @@ CURLcode Curl_async_is_resolved(struct Curl_easy *data,
 
   if(!ares->num_pending) {
     /* all c-ares operations done, what is the result to report? */
-    Curl_resolv_unlink(data, &data->state.async.dns);
+    Curl_dns_entry_unlink(data, &data->state.async.dns);
     data->state.async.done = TRUE;
     result = ares->result;
     if(ares->ares_status == ARES_SUCCESS && !result) {
       data->state.async.dns =
-        Curl_dnscache_mk_entry(data, &ares->temp_ai,
-                               data->state.async.hostname, 0,
-                               data->state.async.port, FALSE);
+        Curl_dns_entry_create(data, &ares->temp_ai,
+                              data->state.async.hostname, 0,
+                              data->state.async.port, FALSE);
       if(!data->state.async.dns) {
         result = CURLE_OUT_OF_MEMORY;
         goto out;
@@ -356,7 +356,7 @@ CURLcode Curl_async_is_resolved(struct Curl_easy *data,
     }
 
     if(result)
-      Curl_resolv_unlink(data, &data->state.async.dns);
+      Curl_dns_entry_unlink(data, &data->state.async.dns);
     *dns = data->state.async.dns;
     CURL_TRC_DNS(data, "is_resolved() result=%d, dns=%sfound",
                  result, *dns ? "" : "not ");
