+++ lib/setopt.c
@@ -1508,8 +1508,8 @@ static CURLcode setopt_pointers(struct Curl_easy *data, CURLoption option,
         data->psl = data->multi ? &data->multi->psl : NULL;
 #endif
       if(data->share->specifier & (1 << CURL_LOCK_DATA_DNS)) {
-        Curl_resolv_unlink(data, &data->state.dns[0]);
-        Curl_resolv_unlink(data, &data->state.dns[1]);
+        Curl_dns_entry_unlink(data, &data->state.dns[0]);
+        Curl_dns_entry_unlink(data, &data->state.dns[1]);
       }
 
       data->share->dirty--;
